Spent most of this time messing with the build system to get the clang lsp to properly identify system headers.
Also, removed most of the pdf stuff. For now I am going to focus on
building a much simpler parser that will walk a pdf and produce the
raw text. It will not be fully what I wanted, but it will be a good
place to start and learn some things.

I added zlib to be able to decode the compressed streams. I don't want
to just work with pdfs I compile by hand and the ones I generated with
lualatex use deflate to compress the streams. This will give me some
practice with that.