Implemented a good portion of the parser.
Tests for the implemented parts pass, though as before they are only very simple tests for well formed input. There are a number of places where EOF is not handled, or where loops and general control flow/process could be improved.

One issue I ran into, the general object parser needs to return PdfObj* objects but some of the pdf objects need specific object types and should hold the obj and not a pointer to it. These are notably the dictionary keys and the stream dictionary object. I tried something to make this work with the names, without duplicating code, and it somehow caused a problem with the switch statement where after creating the name instead of continuing in the block it would run the first block as if we were parsing a null (which should require seeing an 'n' first and is above the name parsing switch) where it would fail because the token it found was just "/". This makes no sense to me unluess the way that I was trying to create the pdf name and store it in a pointer is wrong. In hindsight I may be able to extract the part where we collect the token into a function rather than the creation of the object. Then I can create the with the result of that function. If that does not work I will just have to leave the duplicated code, but this bug is very strange to me and I will have to ask someone for insight and maybe reproduce the problem with a smaller simpler example.