Fixed the issue with duplicated code for parsing names.
The previous commit mentioned an error with the way I was parsing the name tokens and the switch statment. I tried what I mentioned there and extracted the part of the logic that parses the text part of the name (assuming the '/' was already parsed) and passes it to the constructor for the object. This makes still makes it pretty easy to generate the two different kinds of objects and does not mess up the switch. I still have no idea what the issue was, but it would seem maybe the switch should be replaced with if statments when there is complex logic and oject initialization needed in the branches.