Got all the parser tests to pass.
As I have mentioned a few times, the tests are not comprehensive, and the TODOs should show that there are lots of things that are not properly covered in the parser yet. Now that I have the logic for most of it implemented I can go back through and make it all more robust and test some edge cases and make sure I catch errors from malformed objects. I can also try and make the process each object is parsed by consistent.